# packages/agile-security/docs/api/swagger/entity-api-swagger.yml
swagger: '2.0'
info:
  title: Agile IDM (Entity) HTTP API
  description: The Agile HTTP API to interact with AGILE IDM through http(s)
  version: 1.0.0
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# the domain of the service
host: localhost:3000
...
schemes:
  - http
# will be prefixed to all paths
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  Oauth2AuthCode:
    type: oauth2
    flow: accessCode
    authorizationUrl: /oauth2/dialog/authorize
    tokenUrl: /oauth/token
  Oauth2Implicit:
    type: oauth2
    flow: implicit
    authorizationUrl: /oauth2/dialog/authorize
  Oauth2ClientCred:
    type: oauth2
    flow: application
    tokenUrl: /oauth2/token
paths:
...
  '/user/{owner}/group/{groupName}':
    parameters:
      -
        name: owner
        in: path
        description: 'Owner of the gruop'
        required: true
        type: string
      -
        name: groupName
        in: path
        description: 'Group name'
        required: true
        type: string
    get:
      security:
      -  Oauth2AuthCode: []
      - Oauth2Implicit: []
      - Oauth2ClientCred: []

      operationId: readGroup
      tags:
        - Group
      description: 'Return a given group by owner and name'
      responses:
        '200':
          description: 'Group read done'
          schema:
            $ref: '#/definitions/Group'
          examples:
            application/json:
              owner: bob!@!local
              group_name: family
              entities:
              -
                id: bob!@!local
                type: /user
                owner: bob!@!local
                user_name: bob
                auth_type: local
              -
                id: fbd142fd2312
                type: /sensor
                owner: alice!@!github
                name: living room humidity
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '404':
          description: 'No Group found'
        '500':
          description: 'Unexpected error'

    delete:
      security:
        -  Oauth2AuthCode: []
        - Oauth2Implicit: []
        - Oauth2ClientCred: []
      operationId: DeleteGroup
      tags:
        - Group
      description: 'Delete a given group.'
      responses:
        '200':
          description: 'Entity deleted successfully'
        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '404':
          description: 'Group not found'
        '500':
          description: 'Unexpected error'
  '/group/':
    post:
      security:
        -  Oauth2AuthCode: []
        - Oauth2Implicit: []
        - Oauth2ClientCred: []
      operationId: storeGroup
      tags:
        - Group
      description: 'Store a group owned by the user for which the Oauth2 token has been obtained'
      parameters:
        -
          name: Group
          in: body
          description: 'The Group to be stored'
          required: true
          schema:
            $ref: '#/definitions/Group'

      responses:
        '200':
          description: 'Entity stored successfully'
          schema:
            $ref: '#/definitions/Group'
          examples:
            application/json:
              owner: bob!@!local
              group_name: family
              entities: []

        '401':
          description: 'Not authenticated'
        '403':
          description: Forbidden
        '409':
          description: 'Conflict. This can happen when Groups already exist for example'
        '500':
          description: 'Unexpected error'
  '/user/{owner}/group/{groupName}/entities/{entityType}/{entityId}':
    parameters:
      -
        name: owner
        in: path
        description: 'Owner of the gruop'
        required: true
        type: string
      -
        name: groupName
        in: path
        description: 'Group name'
        required: true
        type: string
      -
        name: entityType
        in: path
        description: 'Entity type'
        required: true
        type: string
      -
        name: entityId
        in: path
        description: 'Entity id. This id is only unique for a certain entity type'
        required: true
        type: string
    post:
      security:
        -  Oauth2AuthCode: []
        - Oauth2Implicit: []
        - Oauth2ClientCred: []
      operationId: addEntityToGroup
      tags:
        - Group
      description: 'Add entity to a group'
      responses:
        '200':
          description: 'Entity added to group  successfully'
          schema:
            $ref: '#/definitions/Group'
          examples:
            application/json:
              owner: bob!@!local
              group_name: family
              entities:
              -
                id: bob!@!local
                type: /user
                owner: bob!@!local
                user_name: bob
                auth_type: local
              -
                id: fbd142fd2312
                type: /sensor
                owner: alice!@!github
                name: living room humidity

        '401':
          description: 'Not authenticated'
        '404':
          description: 'Entity owner or group not found'
        '403':
          description: Forbidden
        '409':
          description: 'Conflict. This can happen when Groups already exist for example'
        '500':
          description: 'Unexpected error'
    delete:
      security:
        - Oauth2AuthCode: []
        - Oauth2Implicit: []
        - Oauth2ClientCred: []
      operationId: deleteEntityToGroup
      tags:
        - Group
      description: 'Delete entity from a group'
      responses:
        '200':
          description: 'Entity removed from group  successfully'
          schema:
            $ref: '#/definitions/Group'
          examples:
            application/json:
              owner: bob!@!local
              group_name: family
              entities:
              -
                id: bob!@!local
                type: /user
                owner: bob!@!local
                user_name: bob
                auth_type: local

        '401':
          description: 'Not authenticated'
        '404':
          description: 'Entity owner or group not found'
        '403':
          description: Forbidden
        '409':
          description: 'Conflict. This can happen when the entity was not in the group.'
        '500':
          description: 'Unexpected error'
definitions:
...
  Entity:
    description: 'An AGILE IDM Entity. This model shows only fields placed in the entity by AGILE. In addition to these fields, each object will have the attributes provided when the entity was created.'
    type: object
    properties:
      id:
        type: string
      type:
        type: string
      owner:
        type: string
  Group:
    description: Group object
    type: object
    properties:
      group_name:
        type: string
      owner:
        type: string
      entities:
        type: array
        items:
          $ref: '#/definitions/Entity'
...